+++ src/linalg.jl
@@ -1,4 +1,4 @@
-export opInverse, opCholesky, opLDL, opHouseholder, opHermitian
+export opInverse, opCholesky, opHouseholder, opLDL, opHermitian
 
 function mulFact!(res, F, v, α, β::T) where {T}
   if β == zero(T)
@@ -67,38 +67,12 @@ argument will perform a cheap hermicity check.
 If M is sparse and real, then only the upper triangle should be stored in order to use 
 [`LDLFactorizations.jl`](https://github.com/JuliaSmoothOptimizers/LDLFactorizations.jl):
 
+    using LDLFactorizations
     triu!(M)
     opLDL(Symmetric(M, :U))
 
 """
-function opLDL(M::AbstractMatrix; check::Bool = false)
-  (m, n) = size(M)
-  m == n || throw(LinearOperatorException("shape mismatch"))
-  if check
-    check_hermitian(M) || throw(LinearOperatorException("matrix is not Hermitian"))
-  end
-  LDL = ldlt(M)
-  prod! = @closure (res, v, α, β) -> mulFact!(res, LDL, v, α, β)
-  tprod! = @closure (res, u, α, β) -> tmulFact!(res, LDL, u, α, β)  # M.' = conj(M)
-  ctprod! = @closure (res, w, α, β) -> mulFact!(res, LDL, w, α, β)
-  S = eltype(LDL)
-  return LinearOperator{S}(m, m, isreal(M), true, prod!, tprod!, ctprod!)
-  #TODO: use iterative refinement.
-end
-
-function opLDL(M::Symmetric{T, SparseMatrixCSC{T, Int}}; check::Bool = false) where {T <: Real}
-  (m, n) = size(M)
-  m == n || throw(LinearOperatorException("shape mismatch"))
-  if check
-    check_hermitian(M) || throw(LinearOperatorException("matrix is not Hermitian"))
-  end
-  LDL = ldl(M)
-  prod! = @closure (res, v) -> ldiv!(res, LDL, v)
-  tprod! = @closure (res, u) -> ldiv!(res, LDL, u)  # M.' = conj(M)
-  ctprod! = @closure (res, w) -> ldiv!(res, LDL, w)
-  S = eltype(LDL)
-  return LinearOperator{S}(m, m, isreal(M), true, prod!, tprod!, ctprod!)
-end
+function opLDL end
 
 function mulHouseholder!(res, h, v, α, β::T) where {T}
   if β == zero(T)
